:zap: refactor for using stack model
